admin command refactor overhaul part 1
